Resctructuring project and extending
